Add recurrence support, new collector, bugfix, ++
First step into a new major version (2.0), as we are breaking backwards
compatibility. This update introduces a new collector class,
`EventsCollector`, and adds support for recurrences. Existing classes
are also improved and expanded, and there are a few bug fixes - most
notably a fix for the single-digit month bug in `EventCalendar`. (E.g.
expected `2016-06-20` but produced `2016-6-20`, which could not be
parsed correctly.)